+++ db/column_family.h
@@ -402,12 +402,12 @@ class ColumnFamilyData {
                                     bool allow_data_in_errors, bool* overlap);
 
   // Get user key range of memtables. Tombstones are counted.
-  Status GetMemtablesUserKeyRange(PinnableSlice& smallest,
-                                  PinnableSlice& largest, bool& found);
+  Status GetMemtablesUserKeyRange(PinnableSlice* smallest,
+                                  PinnableSlice* largest, bool* found);
 
   // Get user key range of all data. Tombstones are counted.
-  Status GetUserKeyRange(PinnableSlice& smallest, PinnableSlice& largest,
-                         bool& found);
+  Status GetUserKeyRange(PinnableSlice* smallest, PinnableSlice* largest,
+                         bool* found);
 
   // A flag to tell a manual compaction is to compact all levels together
   // instead of a specific level.
